+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+### 2.0.1
+
+* Change order creation endpoint to POST
+
 ### 2.0.0
 
 * Major rewrite of the dispersal api. This is now namespaced to Dispersal.
@@ -7,7 +11,6 @@
 * Add endpoint to report product queue state
 * Ensure that API sends body in JSON format
 * Allow dispersal vendor to report dispersal process completed
-* Change order creation endpoint to POST
 
 ### 1.1.0